Prevent inactivity-kill of the app if user does not visit the task for a while
As described in the Android developer documentation, Android "clears a task"
"if the user hasn't visited the task for a certain amount of time,
such as 30 minutes.":
http://developer.android.com/guide/topics/manifest/activity-element.html#always

And it says:
However, when this attribute is "true", users will always return to the task in
its last state, regardless of how they get there. This is useful, for example,
in an application like the web browser where there is a lot of state (such as
multiple open tabs) that users would not like to lose.

The same applies to the SDL applications in this repository, so we want this:
android:alwaysRetainTaskState="true".