[opentyrian] JE_pauseGame(): #ifdef ANDROID: Pause audio playback
JE_pauseGame() is used in-Game to pause a running game. On mobile
devices which do multi-tasking like Android, we likely want to stop
Audio playback when the game is paused. Therefore (for now within
the audio playback task completely and block all threads to consume
no CPU cycles while paused. That's to be developed in a future commit.